requirements.txt
PyQt6>=6.7
Pillow>=10.4

